#86 Wrong usage of ViewHolder pattern in AppList causing duplicated view states
